+++ app.js
@@ -6,7 +6,7 @@ var logger = require('morgan');
 
 var indexRouter = require('./routes/index');
 var usersRouter = require('./routes/users');
-const memberRouter=require('./routes/createMemberApi');
+const memberRouter=require('./routes/memberApi');
 
 var app = express();
 
@@ -22,7 +22,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 
 app.use('/', indexRouter);
 app.use('/users', usersRouter);
-app.use('/createMember',memberRouter);
+app.use('/memberApi',memberRouter);
 app.use('/public',express.static('public'));
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
