+++ models/memberModels.js
@@ -1,17 +1,17 @@
 const mongoose = require('mongoose');
-const url = 'mongodb://localhost:27017/project_web';
-mongoose.connect(url, {useNewUrlParser: true, useUnifiedTopology: true});
+const url = 'mongodb://localhost:27017/Project_Perfume';
+mongoose.connect(url, { useNewUrlParser: true, useUnifiedTopology: true });
 const db = mongoose.connection;
 db.on('error', console.error.bind(console, 'connection error:'));
-db.once('open', function() {
-console.log("連線成功");
+db.once('open', function () {
+    console.log("連線成功");
 });
 const memberSchema = new mongoose.Schema({
-userRole:String,
-userName: String,
-account: String,
-password: String
+    userRole: String,
+    userName: String,
+    account: String,
+    password: String
 });
-memberSchema.set('collection', 'member');
-const model = mongoose.model('member', memberSchema);
+memberSchema.set('collection', 'Accounts');
+const model = mongoose.model('Accounts', memberSchema);
 module.exports = model;
